dc++ patchar. några kraschbuggar fixade =) /ctopic och /topic snyggat upp lite varningar

+++ trunk/windows/FulEditCtrl.h
@@ -49,7 +49,7 @@ private:
 	CHARFORMAT2 selFormat;
 	char*		findBuffer;
 	int			curFindPos;
-	const int	findBufferSize;
+	const WORD	findBufferSize;
     static UINT	WM_FINDREPLACE;
 	StringList	urls;
